feat: 상품 이미지 등록 컴포넌트 분리
